requirements.txt
# Scientific Libraries
numpy
scipy

# Image Processing Library
pillow

# Emotion Recognition
opencv-python
deepface

# Gesture Recognition
mediapipe; platform_system == "Windows"
mediapipe-rpi4; platform_system == "Linux"

# Library for 1.3" LCD Circular Displays
spidev; platform_system == "Linux"
st7789; platform_system == "Linux"